+++ diagflow_45/.packages
@@ -3,7 +3,7 @@
 # 
 # For more info see: https://dart.dev/go/dot-packages-deprecation
 # 
-# Generated by pub on 2021-07-24 22:28:55.380472.
+# Generated by pub on 2021-07-24 22:36:07.665885.
 async:file:///C:/Padikkan_Flutter/flutter/.pub-cache/hosted/pub.dartlang.org/async-2.6.1/lib/
 boolean_selector:file:///C:/Padikkan_Flutter/flutter/.pub-cache/hosted/pub.dartlang.org/boolean_selector-2.1.0/lib/
 bubble:file:///C:/Padikkan_Flutter/flutter/.pub-cache/hosted/pub.dartlang.org/bubble-1.2.1/lib/
