+++ diagflow_45/lib/main.dart
@@ -1,254 +1,21 @@
-// import 'package:diagflow_45/screens/chatbot.dart';
-// import 'package:flutter/material.dart';
-
-// void main() => runApp(MyApp());
-
-// class MyApp extends StatelessWidget {
-//   @override
-//   Widget build(BuildContext context) {
-//     return MaterialApp(
-//       home: SafeArea(
-//         child: Scaffold(
-//           body: ChatBot(),
-//         ),
-//       ),
-//     );
-//   }
-// }
-
-import 'package:bubble/bubble.dart';
+import 'package:diagflow_45/screens/chatbot.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_dialogflow/dialogflow_v2.dart';
-import 'package:intl/intl.dart';
 
 void main() {
   runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Flutter Demo',
       theme: ThemeData(
         brightness: Brightness.dark,
-        // This is the theme of your application.
-        //
-        // Try running your application with "flutter run". You'll see the
-        // application has a blue toolbar. Then, without quitting the app, try
-        // changing the primarySwatch below to Colors.green and then invoke
-        // "hot reload" (press "r" in the console where you ran "flutter run",
-        // or simply save your changes to "hot reload" in a Flutter IDE).
-        // Notice that the counter didn't reset back to zero; the application
-        // is not restarted.
         primarySwatch: Colors.blue,
-        // This makes the visual density adapt to the platform that you run
-        // the app on. For desktop platforms, the controls will be smaller and
-        // closer together (more dense) than on mobile platforms.
         visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
       debugShowCheckedModeBanner: false,
-      home: MyHomePage(title: 'Flutter Demo Home Page'),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  MyHomePage({
-    this.title,
-  });
-
-// This widget is the home page of your application. It is stateful, meaning
-// that it has a State object (defined below) that contains fields that affect
-// how it looks.
-
-// This class is the configuration for the state. It holds the values (in this
-// case the title) provided by the parent (in this case the App widget) and
-// used by the build method of the State. Fields in a Widget subclass are
-// always marked "final".
-
-  final String title;
-
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  void response(query) async {
-    AuthGoogle authGoogle =
-        await AuthGoogle(fileJson: "assets/service.json").build();
-    Dialogflow dialogflow =
-        Dialogflow(authGoogle: authGoogle, language: Language.english);
-    AIResponse aiResponse = await dialogflow.detectIntent(query);
-    setState(() {
-      messsages.insert(0, {
-        "data": 0,
-        "message": aiResponse.getListMessage()[0]["text"]["text"][0].toString()
-      });
-    });
-
-    print(aiResponse.getListMessage()[0]["text"]["text"][0].toString());
-  }
-
-  final messageInsert = TextEditingController();
-  List<Map> messsages = <Map>[];
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(
-          "Chat bot",
-        ),
-      ),
-      body: Container(
-        child: Column(
-          children: <Widget>[
-            Container(
-              padding: EdgeInsets.only(top: 15, bottom: 10),
-              child: Text(
-                "Today, ${DateFormat("Hm").format(DateTime.now())}",
-                style: TextStyle(fontSize: 20),
-              ),
-            ),
-            Flexible(
-                child: ListView.builder(
-                    reverse: true,
-                    itemCount: messsages.length,
-                    itemBuilder: (context, index) => chat(
-                        messsages[index]["message"].toString(),
-                        messsages[index]["data"]))),
-            SizedBox(
-              height: 20,
-            ),
-            Divider(
-              height: 5.0,
-              color: Colors.greenAccent,
-            ),
-            Container(
-              child: ListTile(
-                leading: IconButton(
-                  icon: Icon(
-                    Icons.camera_alt,
-                    color: Colors.greenAccent,
-                    size: 35,
-                  ),
-                  onPressed: () {},
-                ),
-                title: Container(
-                  height: 35,
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.all(Radius.circular(15)),
-                    color: Color.fromRGBO(220, 220, 220, 1),
-                  ),
-                  padding: EdgeInsets.only(left: 15),
-                  child: TextFormField(
-                    controller: messageInsert,
-                    decoration: InputDecoration(
-                      hintText: "Enter a Message...",
-                      hintStyle: TextStyle(color: Colors.black26),
-                      border: InputBorder.none,
-                      focusedBorder: InputBorder.none,
-                      enabledBorder: InputBorder.none,
-                      errorBorder: InputBorder.none,
-                      disabledBorder: InputBorder.none,
-                    ),
-                    style: TextStyle(fontSize: 16, color: Colors.black),
-                    onChanged: (value) {},
-                  ),
-                ),
-                trailing: IconButton(
-                    icon: Icon(
-                      Icons.send,
-                      size: 30.0,
-                      color: Colors.greenAccent,
-                    ),
-                    onPressed: () {
-                      if (messageInsert.text.isEmpty) {
-                        print("empty message");
-                      } else {
-                        setState(() {
-                          messsages.insert(
-                              0, {"data": 1, "message": messageInsert.text});
-                        });
-                        response(messageInsert.text);
-                        messageInsert.clear();
-                      }
-                      FocusScopeNode currentFocus = FocusScope.of(context);
-                      if (!currentFocus.hasPrimaryFocus) {
-                        currentFocus.unfocus();
-                      }
-                    }),
-              ),
-            ),
-            SizedBox(
-              height: 15.0,
-            )
-          ],
-        ),
-      ),
-    );
-  }
-
-  //for better one i have use the bubble package check out the pubspec.yaml
-
-  Widget chat(String message, int data) {
-    return Container(
-      padding: EdgeInsets.only(left: 20, right: 20),
-      child: Row(
-        mainAxisAlignment:
-            data == 1 ? MainAxisAlignment.end : MainAxisAlignment.start,
-        children: [
-          data == 0
-              ? Container(
-                  height: 60,
-                  width: 60,
-                  child: CircleAvatar(
-                    backgroundImage: AssetImage("assets/robot.jpg"),
-                  ),
-                )
-              : Container(),
-          Padding(
-            padding: EdgeInsets.all(10.0),
-            child: Bubble(
-                radius: Radius.circular(15.0),
-                color: data == 0
-                    ? Color.fromRGBO(23, 157, 139, 1)
-                    : Colors.orangeAccent,
-                elevation: 0.0,
-                child: Padding(
-                  padding: EdgeInsets.all(2.0),
-                  child: Row(
-                    mainAxisSize: MainAxisSize.min,
-                    children: <Widget>[
-                      SizedBox(
-                        width: 10.0,
-                      ),
-                      Flexible(
-                          child: Container(
-                        constraints: BoxConstraints(maxWidth: 200),
-                        child: Text(
-                          message,
-                          style: TextStyle(
-                              color: Colors.white, fontWeight: FontWeight.bold),
-                        ),
-                      ))
-                    ],
-                  ),
-                )),
-          ),
-          data == 1
-              ? Container(
-                  height: 60,
-                  width: 60,
-                  child: CircleAvatar(
-                    backgroundImage: AssetImage("assets/default.jpg"),
-                  ),
-                )
-              : Container(),
-        ],
-      ),
+      home: ChatBot(),
     );
   }
 }
