+++ diagflow_45/pubspec.yaml
@@ -48,8 +48,12 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
+  assets:
+    - assets/json/dialogflow-45.json
+    - assets/images/camera.png
+    - assets/images/default.jpg
+    - assets/images/default.jpg
+    - assets/images/robot.jpg
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
